tests: Refactor duplicate code for checking secure Windows permissions
